python_files/requirements.txt
flask>=3.0.0
langchain>=0.3.0
langchain-groq>=0.2.0
groq>=0.11.0
python-dotenv>=1.0.0